test 2

+++ src/common.c
@@ -2,22 +2,22 @@
 
 static bool read_callback ( pb_istream_t *stream, uint8_t* buf, size_t count )
 {
-	int fd = (intptr_t)stream->state;
-	int result;
+	int fd = (intptr_t) stream->state;
+	ssize_t result;
 
-	result = recv(fd, buf, count, MSG_WAITALL);
+	result = recv ( fd, buf, count, MSG_WAITALL );
 
 	if ( result == 0 )
 		stream->bytes_left = 0;
 
-	return result == count;
+	return (unsigned) result == count;
 }
 
 static bool write_callback ( pb_ostream_t *stream, const uint8_t *buf, size_t count )
 {
 	int fd = (intptr_t)stream->state;
 
-	return send ( fd, buf, count, 0 ) == count;
+	return (unsigned) send ( fd, buf, count, 0 ) == count;
 }
 
 pb_ostream_t pb_ostream_from_socket ( int fd )
